+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Changed
 - Rename `LICENSE.md` to `LICENSE`
+- Upgrade gulp to version 4
 
 ## 0.4.0 - 2018-09-27