+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Add PHP `7.3` to travis-ci build matrix
 - Set up JS module bundling with [rollup js](https://rollupjs.org)
 - Set up [broser-sync](https://www.browsersync.io)
+- Add `.gitattributes`
 
 ### Changed
 - Rename theme back to `jentil-theme`