+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 - Add `.security.txt`
-- Add PHP `7.3` to travis-ci build matrix
+- Add suport for PHP 7.3 and 7.4
 - Set up JS module bundling with [rollup js](https://rollupjs.org)
 - Set up [broser-sync](https://www.browsersync.io)
 - Add `.gitattributes`